requirements.txt
Django==1.9.7
djangorestframework==3.3.3
MySQL-python==1.2.5
django-rest-swagger==0.3.7
